requirements.txt
# Core ML & AI
openai>=1.12.0
anthropic>=0.18.0
sentence-transformers>=2.3.1
torch>=2.1.0

# RAG & Vector DB
chromadb>=0.4.22
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.23
rank-bm25>=0.2.2

# Multi-Agent Framework
langgraph>=0.0.25

# Data Collection
arxiv>=2.1.0
scholarly
requests>=2.31.0
pydantic>=2.5.0
beautifulsoup4>=4.12.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0  # Statistical tests (Wilcoxon, binomial, Cohen's d)

# PDF Processing
pypdf>=4.0.0
pdfminer.six>=20221105

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0

# UI
gradio>=4.16.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
ipykernel>=6.29.0
black>=24.1.0
ruff>=0.1.0